Adding safety checks so we only create the admin user and alter the cassandra user if we need to.

+++ src/start.sh
@@ -52,18 +52,38 @@ if [ -f $passwordFile ]; then
 	sleep 10s
 
 	configFile=$CASSANDRADIR/userConfig.cql
+	adminPasswordFile=$CASSANDRADIR/cassandraPasswords.txt
+	
+	if [ ! -f $adminPasswordFile ]; then
+		randomPassword=`cat /dev/urandom | tr -dc A-Za-z0-9 | head -c 32`
+
+		echo "ALTER USER cassandra WITH PASSWORD '$randomPassword' SUPERUSER;" > $CASSANDRADIR/alterDefaultUser.cql
+		echo "cassandra:$randomPassword" > $adminPasswordFile
+
+		adminPassword=`cat /dev/urandom | tr -dc A-Za-z0-9 | head -c 32`
+		echo "admin:$adminPassword" >> $adminPasswordFile
+		echo "CREATE USER IF NOT EXISTS 'admin' WITH PASSWORD '$adminPassword' SUPERUSER;" > $configFile
+		cqlsh -u cassandra -p cassandra -f $configFile
+		cqlsh -u admin -p $adminPassword -f $CASSANDRADIR/alterDefaultUser.cql
+		echo "" > $configFile #clear file
+	else
+		IFS=$'\n'
+		while read line
+		do
+				actualCreds+=("$line")
+		done < $adminPasswordFile
+
+		IFS=$':'
+		for ((i=0; i < ${#actualCreds[*]}; i++))
+		do
+				read -r user pass <<< "${actualCreds[i]}"
+				if [ "$user" -eq "admin" ]; then
+					adminPassword="$pass"
+				fi
+		done
+	fi
 
 	IFS=$'\n'
-	randomPassword=`cat /dev/urandom | tr -dc A-Za-z0-9 | head -c 32`
-
-	echo "ALTER USER cassandra WITH PASSWORD '$randomPassword' SUPERUSER;" > $CASSANDRADIR/alterDefaultUser.cql
-	echo "cassandra:$randomPassword" > /root/cassandraPasswords.txt
-
-	randomPasswordAdmin=`cat /dev/urandom | tr -dc A-Za-z0-9 | head -c 32`
-	echo "admin:$randomPasswordAdmin" >> /root/cassandraPasswords.txt
-
-	echo "CREATE USER IF NOT EXISTS 'admin' WITH PASSWORD '$randomPasswordAdmin' SUPERUSER;" > $configFile
-
 	while read line
 	do
 			array+=("$line")
@@ -76,7 +96,5 @@ if [ -f $passwordFile ]; then
 			echo "CREATE USER IF NOT EXISTS '$user' WITH PASSWORD '$pass' SUPERUSER;" >> $configFile
 	done
 
-	cqlsh -u cassandra -p cassandra -f $configFile
-	cqlsh -u admin -p $randomPasswordAdmin -f $CASSANDRADIR/alterDefaultUser.cql
-
+	cqlsh -u admin -p $adminPassword -f $configFile
 fi
